Making intersection functions
